Major release 2.0: Added configuration menu
On first run (no key in localStorage) you get prompted with
the configuration menu. You can then open the menu again by
clicking on the 'Config' label button on the site page.

Configuration menu css is loaded via @resource.

Other minior changes and patches also come with this release.
Among the few changes is the @run-at setting, which now is
'document-start'. The script does some things at document start
and then waits for the DOM to load completely so it can finish
the setup. This is much more consistent with the main page load.